% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BootEst.MO.SA}
\alias{BootEst.MO.SA}
\title{Bootstrap confidence intervals for the bias-corrected multiply robust estimator under violations of monotonicity}
\usage{
BootEst.MO.SA(
  times,
  propensity.model,
  principal.model,
  censor.model,
  failure.model,
  data,
  zeta,
  estimand = c("NACE", "CACE", "AACE", "DACE"),
  bootstrap = 50
)
}
\arguments{
\item{times}{a vector of time when the principal survival causal effects (PSCEs) are of interest}

\item{propensity.model}{propensity score model}

\item{principal.model}{principal score model}

\item{censor.model}{censoring model}

\item{failure.model}{outcome model}

\item{data}{dataset}

\item{zeta}{the sensitivity parameter zeta}

\item{estimand}{the estimands of interest}

\item{bootstrap}{number of bootstrap replications}
}
\value{
The bootstrap confidence intervals
}
\description{
Bootstrap confidence intervals for the bias-corrected multiply robust estimator under violations of monotonicity
}
