% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BootEst.PI.SA}
\alias{BootEst.PI.SA}
\title{Bootstrap confidence intervals for the multiply robust estimator}
\usage{
BootEst.PI.SA(
  times,
  propensity.model,
  principal.model,
  censor.model,
  failure.model,
  data,
  xi0,
  xi1,
  eta0,
  eta1,
  estimand = c("NACE", "CACE", "AACE"),
  bootstrap
)
}
\arguments{
\item{times}{a vector of time when the principal survival causal effects (PSCEs) are of interest}

\item{propensity.model}{propensity score model}

\item{principal.model}{principal score model}

\item{censor.model}{censoring model}

\item{failure.model}{outcome model}

\item{data}{dataset}

\item{xi0}{sensitivity parameter xi_0}

\item{xi1}{sensitivity parameter xi_1}

\item{eta0}{sensitivity parameter eta_0}

\item{eta1}{sensitivity parameter eta_1}

\item{estimand}{the estimands of interest}

\item{bootstrap}{number of bootstrap replications}
}
\value{
The bootstrap confidence intervals
}
\description{
Bootstrap confidence intervals for the multiply robust estimator
}
