% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mrPStrata}
\alias{mrPStrata}
\title{Multiply robust estimator for calculating the principal survival causal effects among always takers, compliers, and never takers}
\usage{
mrPStrata(
  times,
  data,
  Xpi_names,
  Xe_names,
  Xc_names,
  Xt_names,
  Z_name,
  S_name,
  U_name,
  delta_name,
  B = 100
)
}
\arguments{
\item{times}{a vector of time when the principal survival causal effects (PSCEs) are of interest}

\item{data}{the dataset}

\item{Xpi_names}{names of the covariates for the propensity score model}

\item{Xe_names}{names of the covariates for the principal score model}

\item{Xc_names}{names of the covariates for the censoring model}

\item{Xt_names}{names of the covariates for the failure outcome model}

\item{Z_name}{name of the treatment assignment status}

\item{S_name}{name of the true treatment receipt status}

\item{U_name}{name of the observed failuture time}

\item{delta_name}{name of the censoring indicator}

\item{B}{number of the bootstrap replications (default 100)}
}
\value{
The PSCE estimates and their 95\% confidence intervals
}
\description{
Multiply robust estimator for calculating the principal survival causal effects among always takers, compliers, and never takers
}
\examples{
# example code
attach(sim_data)
res = mrPStrata(times=c(1,2,3,4,5,6,7,8),
                data = sim_data,
                Xpi_names = c("X1","X2","X3","X4","X5"),
                Xe_names = c("X1","X2","X3","X4","X5"),
                Xc_names = c("X1","X2","X3","X4","X5"),
                Xt_names = c("X1","X2","X3","X4","X5"),
                Z_name = "z",
                S_name = "s",
                U_name ="U",
                delta_name = "delta",
                B=50)
print(res)
}
