% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mrPStrata_MO_SA}
\alias{mrPStrata_MO_SA}
\title{Bias-corrected multiply robust estimator of the PSCE under violation of monotonicity}
\usage{
mrPStrata_MO_SA(
  times,
  data,
  Xpi_names,
  Xe_names,
  Xc_names,
  Xt_names,
  Z_name,
  S_name,
  U_name,
  delta_name,
  zeta = 0.01,
  B = 100
)
}
\arguments{
\item{times}{a vector of time when the principal survival causal effects (PSCEs) are of interest}

\item{data}{the dataset}

\item{Xpi_names}{names of the covariates for the propensity score model}

\item{Xe_names}{names of the covariates for the principal score model}

\item{Xc_names}{names of the covariates for the censoring model}

\item{Xt_names}{names of the covariates for the outcome model}

\item{Z_name}{name of the treatment assignment status}

\item{S_name}{name of the true treatment receipt status}

\item{U_name}{name of the observed failuture time}

\item{delta_name}{name of the censoring indicator}

\item{zeta}{sensitivity parameter zeta}

\item{B}{number of the bootstrap replications (default 100)}
}
\value{
The PSCE estimates and their 95\% confidence intervals
}
\description{
Bias-corrected multiply robust estimator of the PSCE under violation of monotonicity
}
