% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mrPStrata_PI_SA}
\alias{mrPStrata_PI_SA}
\title{Bias-corrected multiply robust estimator of the PSCE under violation of the principal ignorability assumption}
\usage{
mrPStrata_PI_SA(
  times,
  data,
  Xpi_names,
  Xe_names,
  Xc_names,
  Xt_names,
  Z_name,
  S_name,
  U_name,
  delta_name,
  xi0 = 0,
  xi1 = 0,
  eta0 = 1,
  eta1 = 1,
  B = 100
)
}
\arguments{
\item{times}{a vector of time when the principal survival causal effects (PSCEs) are of interest}

\item{data}{the dataset}

\item{Xpi_names}{names of the covariates for the propensity score model}

\item{Xe_names}{names of the covariates for the principal score model}

\item{Xc_names}{names of the covariates for the censoring model}

\item{Xt_names}{names of the covariates for the outcome model}

\item{Z_name}{name of the treatment assignment status}

\item{S_name}{name of the true treatment receipt status}

\item{U_name}{name of the observed failuture time}

\item{delta_name}{name of the censoring indicator}

\item{xi0}{sensitivity parameter xi_0}

\item{xi1}{sensitivity parameter xi_1}

\item{eta0}{sensitivity parameter eta_0}

\item{eta1}{sensitivity parameter eta_1}

\item{B}{number of the bootstrap replications (default 100)}
}
\value{
The PSCE estimates and their 95\% confidence intervals
}
\description{
Bias-corrected multiply robust estimator of the PSCE under violation of the principal ignorability assumption
}
