% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.psce}
\alias{plot.psce}
\title{Plot of the PSCEs and their associated 95\% pointwise confidence intervals}
\usage{
\method{plot}{psce}(res)
}
\arguments{
\item{res}{an output from mrPStrata}
}
\value{
The PSCE point estimates and 95\% pointwise confidence intervals
}
\description{
Plot of the PSCEs and their associated 95\% pointwise confidence intervals
}
