% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PointEst.MO.SA}
\alias{PointEst.MO.SA}
\title{Point estimation for the multiply robust estimator under violation of monotonicity assumption}
\usage{
PointEst.MO.SA(
  times,
  propensity.model,
  principal.model,
  censor.model,
  failure.model,
  data,
  zeta,
  estimand = c("NACE", "CACE", "AACE", "DACE")
)
}
\arguments{
\item{times}{a vector of time when the principal survival causal effects (PSCEs) are of interest}

\item{propensity.model}{propensity score model}

\item{principal.model}{principal score model}

\item{censor.model}{censoring model}

\item{failure.model}{outcome model}

\item{data}{dataset}

\item{zeta}{the sensitivity parameter zeta}

\item{estimand}{the estimands of interest}
}
\value{
The PSCE point estimates
}
\description{
Point estimation for the multiply robust estimator under violation of monotonicity assumption
}
